context/requirements.txt
flask>=2.3.0
pillow>=9.0.0
requests>=2.25.0